move ReleaseChannel to shared module
